commit the stuff as it was
seems like setSensorPhase was the fixed the backwards thing.
docs say do not setInverted in closedLoop.

myRunIndex: when the shoot code turned the index on, the ball would go back down.  I'm not sure why.

alighn was commented out because it seemed to be wanting to align even when angle was still set at close.